Move http.ResponseWriter.WriteHeader calls later
unit tests shown that when calling twice WriteHeader, the actual
result returned to the client is the one brought by the first call.

For this reason, for leveraging the improvements made with #19,
we need to ensure that WriteHeader is not called once by
the handler (before panicking) and once again (recovering panic).

Therefore, this change moves WriteHeader to a later point than
the high panick risk code.